+++ packages/config/src/index.ts
@@ -1,6 +1,5 @@
 /* eslint-disable import/export */
 
-export * from './Modules';
 export * from './Config';
 export * from './Config.types';
 export * from './Project';
